Agregados tests e informacion a references

+++ src/main/java/es/uvigo/esei/tfg/repodroid/store/json/JSONStorer.java
@@ -1,4 +1,3 @@
-
 package es.uvigo.esei.tfg.repodroid.store.json;
 
 import com.fasterxml.jackson.databind.DeserializationFeature;
@@ -18,27 +17,33 @@ public class JSONStorer implements Storer {
     ObjectMapper mapper;
     String storerDirectory;
     private Logger logger;
-    
+
     @Override
     public void initialize(String basePath, Logger l) {
-       this.logger = l;
-       this.logger.log(Level.INFO, "Initializing json storer..."); 
-       this.mapper = new ObjectMapper();
-       this.storerDirectory = basePath;
-       File storerDir = new File(basePath);
-       storerDir.mkdirs();
+        if (basePath == null) {
+            throw new IllegalArgumentException("The base path is null");
+        }
+        this.logger = l;
+        this.logger.log(Level.INFO, "Initializing json storer...");
+        this.mapper = new ObjectMapper();
+        this.storerDirectory = basePath;
+        File storerDir = new File(basePath);
+        storerDir.mkdirs();
     }
 
     @Override
     public void close() {
-        this.logger.log(Level.INFO, "Terminating json storer..."); 
+        this.logger.log(Level.INFO, "Terminating json storer...");
     }
 
     @Override
     public void storeSample(Sample sample) {
-        this.logger.log(Level.INFO, "Storing a sample as a json file..."); 
+        if (sample == null) {
+            throw new IllegalArgumentException("The sample is null");
+        }
+        this.logger.log(Level.INFO, "Storing a sample as a json file...");
         try {
-            File jsonSample = new File(this.storerDirectory+File.separator+sample.getId()+".json");
+            File jsonSample = new File(this.storerDirectory + File.separator + sample.getId() + ".json");
             mapper.writeValue(jsonSample, sample);
         } catch (IOException ex) {
             this.logger.log(Level.SEVERE, null, ex);
@@ -47,12 +52,14 @@ public class JSONStorer implements Storer {
 
     @Override
     public Sample retrieveSample(String sampleID) {
-       this.logger.log(Level.INFO, "Retrieving a sample from a json file..."); 
-       Sample sample = new Sample();
+        if (sampleID == null) {
+            throw new IllegalArgumentException("The ID is null");
+        }
+        this.logger.log(Level.INFO, "Retrieving a sample from a json file...");
+        Sample sample = new Sample();
         try {
-            //CORRIGE QUE FALLE PQ NO ENCUENTRA ANALYSISNAME. VER MEJOR FORMA
             mapper.configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false);
-            sample = this.mapper.readValue(new File(this.storerDirectory+File.separator+sampleID+".json"), Sample.class);
+            sample = this.mapper.readValue(new File(this.storerDirectory + File.separator + sampleID + ".json"), Sample.class);
         } catch (IOException ex) {
             this.logger.log(Level.SEVERE, null, ex);
         }
@@ -60,24 +67,29 @@ public class JSONStorer implements Storer {
     }
 
     @Override
-    public void removeSample(String sampleID){
-       this.logger.log(Level.INFO, "Deleting a json file..."); 
-       try {
-            Files.delete(Paths.get(this.storerDirectory+File.separator+sampleID+".json"));
-       }catch (NoSuchFileException x) {
-            System.err.format("%s: no such" + " file or directory%n", this.storerDirectory+File.separator+sampleID+".json");
-       }catch (IOException x) {
+    public void removeSample(String sampleID) {
+        if (sampleID == null) {
+            throw new IllegalArgumentException("The ID is null");
+        }
+        this.logger.log(Level.INFO, "Deleting a json file...");
+        try {
+            Files.delete(Paths.get(this.storerDirectory + File.separator + sampleID + ".json"));
+        } catch (NoSuchFileException x) {
+            System.err.format("%s: no such" + " file or directory%n", this.storerDirectory + File.separator + sampleID + ".json");
+        } catch (IOException x) {
             // File permission problems are caught here.
-        System.err.println(x);
-}
+            System.err.println(x);
+        }
     }
 
-    //ELIMINAMOS Y VOLVEMOS A CREAR EL JSON¿?
     @Override
     public void updateSample(String sampleID, Sample sample) {
-        this.logger.log(Level.INFO, "Updating a json file..."); 
+        if (sampleID == null || sample == null) {
+            throw new IllegalArgumentException("The parameters can't be null");
+        }
+        this.logger.log(Level.INFO, "Updating a json file...");
         this.removeSample(sampleID);
         this.storeSample(sample);
     }
-    
+
 }
